Extract LDAPSchoolSyncer
